backend/requirements.txt
# Updated requirements.txt for Python 3.10+ (including 3.13 on Windows)

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Machine Learning
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.1.0
joblib>=1.3.0

# NLP
nltk>=3.8.1

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.9.0

# Visualization (optional)
matplotlib>=3.8.0
seaborn>=0.13.0
